Files with shape pruning processor and weights flops calculator was renamed
